* code refactoring * added a check whether new token is already in use (not optimal way so far)
