+++ backend/src/Shorty.API/Features/Urls/UrlRepository.cs
@@ -1,19 +1,30 @@
 ﻿using Dapper;
 using Microsoft.Extensions.Caching.Distributed;
 using Npgsql;
+using System;
 
-namespace Shorty.API.Services;
+namespace Shorty.API.Features.Urls;
+
+public interface IUrlRepository
+{
+    Task<string> SaveAsync(string url, CancellationToken cancellationToken = default);
+    Task<string?> GetAsync(string token, CancellationToken cancellationToken = default);
+}
 
 internal sealed class UrlRepository(IDistributedCache cache, NpgsqlConnection db) : IUrlRepository
 {
     public async Task<string> SaveAsync(string url, CancellationToken cancellationToken = default)
     {
-        var value = await GetByUrlAsync(url, cancellationToken);
+        var value = await GetTokenByUrlAsync(url, cancellationToken);
         if (string.IsNullOrEmpty(value))
         {
             var token = new ShortUrlToken();
-            value     = token.GetValue();
-
+            do
+            {
+                value = token.GetValue();
+            }
+            while (await CheckIfTokenIsInUseAsync(value, cancellationToken));
+            
             const string sql = """
                 INSERT INTO shorty_urls (url, token)
                 VALUES (@url, @value)
@@ -27,7 +38,7 @@ internal sealed class UrlRepository(IDistributedCache cache, NpgsqlConnection db
         return value;
     }
 
-    private async Task<string?> GetByUrlAsync(string url, CancellationToken cancellationToken)
+    private async Task<string?> GetTokenByUrlAsync(string url, CancellationToken cancellationToken)
     {
         const string sql = """
             SELECT token 
@@ -38,6 +49,19 @@ internal sealed class UrlRepository(IDistributedCache cache, NpgsqlConnection db
 
         var token = await db.QueryFirstOrDefaultAsync<string?>(sql, new { url });
         return token;
+    }
+
+    private async Task<bool> CheckIfTokenIsInUseAsync(string token, CancellationToken cancellationToken)
+    {
+        const string sql =
+            """
+            SELECT EXISTS
+            (
+                SELECT 1 FROM shorty_urls WHERE token = @token
+            )
+            """;
+
+        return await db.QueryFirstOrDefaultAsync<bool>(sql, new { token });
     }
 
     public async Task<string?> GetAsync(string token, CancellationToken cancellationToken = default)
@@ -55,7 +79,7 @@ internal sealed class UrlRepository(IDistributedCache cache, NpgsqlConnection db
             url = await db.QueryFirstOrDefaultAsync<string>(sql, new { token });
 
             if (!string.IsNullOrEmpty(url))
-            { 
+            {
                 await SaveToCacheAsync(token, url, cancellationToken);
             }
         }
@@ -73,7 +97,7 @@ internal sealed class UrlRepository(IDistributedCache cache, NpgsqlConnection db
 
     private static DistributedCacheEntryOptions CreateDefaultOptions() => new DistributedCacheEntryOptions
     {
-        SlidingExpiration               = TimeSpan.FromHours(5),
+        SlidingExpiration = TimeSpan.FromHours(5),
         AbsoluteExpirationRelativeToNow = TimeSpan.FromDays(30)
     };
 }
